add some flow diagrams, update intent diagram, add responseBuilderBase
