+++ doc/architectual_diagram.drawio
