+++ doc/dialog_intent_map.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2021-12-15T10:10:06.693Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.1.8 Chrome/87.0.4280.88 Electron/11.1.1 Safari/537.36" etag="R3nawi9rh_VvObd5uYlh" version="14.1.8" type="device"><diagram id="Ehb2cejAjsyU7AzATvCr" name="Seite-1">5Vpdc5s4FP01PIaxJSPgceNNu52dpjvJtpntS0aAMGpkRIUaO/vrKzDYRpI7aYshsWc8HnQlPnTO1blXFxw4X67fClxk73lCmAMmydqBfzoATGcAONVvkjxtLEFrWAiaNIN2hlv6P2mMk8b6jSak7AyUnDNJi64x5nlOYtmxYSH4qjss5ax71wIviGG4jTEzrXc0kVk7C39n/4vQRdbeeYrCTc8St4ObmZQZTvhqzwSvHDgXnMvN0XI9J6wCr8Vlc96bA73bBxMkl885YXGLyMXy/h/2+THl5HIFPlxcXTRXecTsWzPhBZFuSfMFI64DEFOXvoyEOlpURwmRmDI3oZjxhdntAFgQUbFdj1PNZuryqcVTPZOiTjUuVxmV5LbAcdWzUt6jbJlcMtWa1tfSp9c+KxGSrPdMzXTfEr4kUjypIU0vChvoG9+DrVOtdkxOW3qyPRZRY8Nli0l76R2+6qCB+Cfgnk1seFcIFZuzla/HghaS8ry6Vi6r2f8yhillbM4ZF/V5ME5IFETKXkrBH8heD0QwhEk/qE9RF/XZxII6sqC+paJ/2K1uXt2pAb3y1vK1Ar7VylZiZ+MDDg75ea0sewoxMVXkVZLgQ01rgvFJmBkkvMsfFQRVuMSlfK3+Huj+7o0PtWdAnXO3FvXqP6YpJYnp6S5RGYrYhtOeWEgwCdLYxgKKAxKl/bAw65LgwWfG1uBoHCCDA5qrBk3cJroOBL9HgmRmgz8AEUSopyirBdnx4fdtkn8Q+k6iY/b+JiU9AKzr+fgAh2enMbrSe89N4I9GQhtqzkNkwommMrZIOyz+ZiavFkGzLT3+KugBU6QrNxodUzNZb326zdJfLpyevse35d3Dwgl/LBFuybgcSifSNAWxVacTFCHvSDoBx/dpc+/TkhBnWOBYVnUqgwNGHslgGp6mBB3gxg8jBWk/MVRLZEA4IDfvr5bXIAXl/ZePT/dz9Hn5iV1bao6nnsf42oZ1yOVhpeCw4p9gFjPVIm44ZBZjhd+sFpx6hIB+2F0CthrloBwcrhacWYCAoEuNP/rysFcSsFoY20r9S6wY6C8+xgcyOLtIGwKNA39sDsyqzQmH2nD20kJtG/vPKNb6E79Lwuixdmqm/GcabJG+FRhdn6bmXmBec/JgJeUAJzfOHDqX84dU1N/TvBk1EOsLwB9/y2vWhGr1sYvPa8A40Eo+/pDvR+wYmyWf6i33S8gMAy0rCYd0yPW/63dfQ3L9990fwZebm/+KIltZajC4oG7OpSsIVpocVR9pDCS8QUzswhsFXvXGoZ/kXBNeaPsqaVAKTN09cQp8TTI8MDYFpiqfOAXA6+71wyHTDysFpmifOAWh9sLqmMVg1dx9X1z37X2lDa++Aw==</diagram></mxfile>
