--- doc/dialog_to_intent.drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2021-12-15T08:37:29.579Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.1.8 Chrome/87.0.4280.88 Electron/11.1.1 Safari/537.36" etag="EriCjT3BRgLiB_i4lTkC" version="14.1.8" type="device"><diagram id="Ehb2cejAjsyU7AzATvCr" name="Seite-1">5Vpdc5s4FP01PIYxyAh43LBJt7PT7E7Sj9m+ZAQIo0ZGLKix019fgYVtJLmTthgS+8VGV0Kgc6Rzry6yQLRcv6lQmb9jKaaWO0vXFvjTcl3H8xzx11iepMXpLIuKpNK2M9yRb1gaZ9L6laS47jXkjFFOyr4xYUWBE96zoapiq36zjNH+U0u0wJrhLkFUt34iKc831sD1d/a/MFnk3ZMdGG5qlqhrLEdS5yhlqz0TuLJAVDHGN1fLdYRpg16Hy+a+6wO12xercMGfc8PiDuKL5f2/9PNjxvDlyv3n4upC9vKI6Fc54AXmdk2KBcW25UIqur6MK3G1aK5SzBGhdkoQZQu92nJBiauG7radKMqh86cOT/FOgjpRuFzlhOO7EiVNzUpMH2HL+ZKKktP2pQ6ve1dccbzeM8nhvsFsiXn1JJrIWhhK6OXkA92kWu2YdDp68j0WobShusOk63qHr7iQEP8E3POZCe8GoXJzt5jrSUVKTljR9FXwZvS/jGFGKI0YZVV7H0hSHAexsNe8Yg94rwZAEIJ0GNQd2Ed9PjOgDg2ob6kYHnbjNG+eJEFvZmv9WgHfaqUEPJhPD7h7aJ63yrKnEDNdRV4lCT5QtCaYnoS5RsLb4lFAsNMbXcCvI1Tz17oSAnUleNOT4GkkFMxu4W9+E5IRnOpE2FjELtXW0Q7EQopwkCUmFmAS4DgbhoV5nwQPPNPrBkfjAGockEIUSGofWAdHgt/DQTo3wR+4MYBwIP+ruN/p4fdNzuAg9L0QSK/9TUoGAFhV+ukBDs9OY1Sl954b2h+NhM7VnIfIhDNFZUyedlz89RhfLAK5YT3+KhgAU6gqN5wcUz2M7+Z0F7+/XDg9dfdvisjHhRP8WCLsmjI+lk5kWeYmRp1OYQy9I+kEmH5O67uijoQkRxVKeJPB0jig+BGPpuFZhuEBbvwwFpAO40OVQMYNR+Tm3dXyxs3c+v7Lh6f7CH5efqQ3hmzkqccxvrJhHXN5GCk4rPgnGMU4iscNx4xijPDr2YJT9xDAD/tLwJS9HJWDw9mCM3MQwO1T40++PMyZBCQWxjaH/xIzBuonkemBDM7O04auwoE/NQd61uaEXW04f2mutvP9Z+Rr/ZnfJ2FyX+voIf+ZOluobgUm1ydH3wtELScPRlIOcHJrRcC6jB6yqj1pcz2pI1YXgD/9llfPCbXqYxaf14BxoKR8/DG/j5gx1lM+Car5S4gMAyUqCceckOv367f/h/jm709/BF9ub/8ry3xlyMGgktgF43aFkdDkuDm+MZLwBgk2C28ceM0Xh2GCc0V4gem80qgU6Lp74hT4imR47tQU6Kp84hS4Xn+vH44Zfhgp0EX7xCkIlQ9Wx0wGi+Lu5HFbt3eAG1x9Bw==</diagram></mxfile>
