+++ doc/flowDiagram/flowDiagram_getAllDetails.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2021-12-15T12:26:05.747Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.1.8 Chrome/87.0.4280.88 Electron/11.1.1 Safari/537.36" etag="DC60Z0o7Ue9iM9tB76cI" version="14.1.8" type="device"><diagram id="ahUZG6pYxQTYNia0-5g8" name="Seite-1">7VpNc9s2EP01mp7qEUiKpo6ObLk5pO3UnckklwxELEWMQUABIcvqry8Agh8SqFiZsSLGyoUmH773LXYfII/CWfF8L/Eq/yAIsFEwJs+j8HYUBAglE/3HINsKiWtgKSlxlVrggf4HDhw7dE0JlDsVlRBM0dUumArOIVU7GJZSbHarZYLtjrrCS/CAhxQzH/1IicorNAmuW/wPoMu8HhnF06qkwHVlt5Iyx0RsOlB4NwpnUghVvRXPM2DGeLVdqnbzA6XNxCRwdUyD9QZmn+/nX/9C0y94cs3Il4L97np5wmztFuwmq7a1BaRYcwKmk/EofLfJqYKHFU5N6UZzrrFcFUx/If1aKikeYSaYkLZ1SOJFPIl1SUYZ6+Dz+fzm7lbjbgIgFTwfXBlq7KUdDUQBSm51FdcgcRbe7n5uWr7iscPyDldR5EDsfGTZ9NyaUb84S36HVQPPqvf0CYxTmIfKzZOAwpSV1idlg5arqvscMKN8aVYhJPE40b60Mq/aT5lYq5d5WYGkem0g20Z/t9BrkICCXRaa7w4NQdRDQ3IqFkKPhT+xJiC86bFvzIwRF5qIeGneIMtMNNGzMo9czxYMSQyedHwzPcS4MCbmi3Jl7Vd1UK4wr7v4mAM35hBrE55wqSzJtOywvC6rOeAOVjJhaopMPwJOmkE1JDtTmTUu0y6F8lQCLsEMsTDUIBI1/gU4zdvu6z7xQljHRKW68q3wD+ZLZ7B4bHoCUH6tqyvddNyYxJZKKDDl1bQaT3cVbCO/mxrRXHfNeHLf3wtMBEOSpX2hLE4TWGSvs1ui8d5umR65W9DJglbkbZdb0fjvBnPrvwZ55DqN6VgmpK43vwx+wvjc0Wzi0fMOp481JzPB9WINRQ8KK/BIsYXfyNRccNiztYN0cFlyw4y2nOXD2JVqnm5cQUEJMcP00rwrHl6BmP1kH/VtnL5sH5yKmPgtSKjreGga6toza5VHgFRrLangB8RTRiUssHkdvmxK9hPB2WVT0uPPvrpppVRr7SMz+jC0VaTyl7UVwYU5H/ZKq/hIaRX26cuDYsdfa0eEockvFfaaWR6Nk8HJsKm3/d6XlT9KG+PMMHlFKbASvqXQtAeu1Y+Mh8Mg8fxarb68unSxNo2HJtbQm7jwQo3sGoxcQ8ffeb2s4IYv21C0nznOrtuQf9/VJyc+2WzBwbJgw5E2ekqz7a5WumxdsH8cGoAuQEffz1RL6QoFQrXktou/DLoGoADexK2Ad6k/gDzj3wv8WzU7+reVlNHCyq5xmtuz1U+QbwK0H5HOn2+Ouyh4b8YubECScmvOqAYhgv+m2pNK9+ze8zvERSWfYF/cDSH5+KfSX8nnAF2nTD76s/2HAVvW+beL8O5/</diagram></mxfile>
