+++ doc/flowDiagram/flowDiagram_getContextDescription.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2021-12-15T13:19:34.535Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.1.8 Chrome/87.0.4280.88 Electron/11.1.1 Safari/537.36" etag="X5QiunXQX1CmWpdBCzEC" version="14.1.8" type="device"><diagram id="8xppoQwRwXuzVS-p-7n9" name="Seite-1">7Vrbbts4EP0aPzYwJVmxHxMn2S0WbYOmQNBHWhxd1pLoUnSc7Ncvh6SulNu0SGDn8mLLwyFFnjMcniE88ZfF/V+CbtJPnEE+8absfuJfTDyPeNNAfaHlwVg8P/CNJREZs16t4Sb7D6xxaq3bjEHVc5Sc5zLb9I0RL0uIZM9GheC7vlvM8/5bNzQBx3AT0dy13mZMpsY6905b+9+QJWn9ZhIuTEtBa2e7kiqljO86Jv9y4i8F59I8FfdLyBG9GhfT72pPazMxAaV8TAf6Mar+PY9PU/HP9e7cD8PV18UHO8odzbd2wXay8qFGQPBtyQAHmU78812aSbjZ0Ahbd4p0ZUtlkatfRD1WUvA1LHnOhe7ts3AVzkLVEmd53rFfXV2dXV4ou50ACAn3e1dGGrxUpAEvQIoH5WI7LGqIbZDN7c9dS1g4tba0Q1YQWCO1QZI0Q7c4qgcL5W/A6jmw3qZU4j7gKphVCKeAIG/MUIzjHBKalRP/ysFfxc0GH1VM5nwrf83BBkSmlgGi7XTdmp4E8LAPOPFcxL1gBPH5cwHuj8RxmCNY1YaW6jnB5zPlsgJaqC8eY9IQNFrnWZmoZyhBJA8aegF0rVniOyoYTlf7ApVbAXqVMlVceVNBywRO1MMnugbrp22swy6VUr2lZriSDf2SigQkdv9S2s6K2Im3HHjoh7WOG8IwL6osJiJsZ7SgZgLfegGl56o7FFzPWKYIwpSX0CKwSwFtD3yLswYapXoCSaqixJvmcAe5Ygkhkztuu1V6PepjJtPaqZ6wABh6ETJ0oyXT898Kx/e09cUVfdfzivS8WSYwt9ew9PrZrUSLHmBc1O0si2PQga3XXw3oqkCFvQKrS5TgGsVYD2KBMaCZkFIR2o2qsUirbSsxtDidn3m7D/IuozCPo7FMHUZzWMVPkyBIs/vrDBE8MkOQxXOliMBJERe8if8d1QEi0bIu1THdbB26QhZUHORMZNLEQk5xG78J4vzw0Kl95vB2brapJmvJS7VaJOhG6m08YEU3/kSilJgT+2BbE82zpERqFHSaEAQ2U0Sd2YYiYwxfM8pzXzU9BTNkoHKCxQgzYzLHey5mQoeZmg19bPxyx7wabuaLPjUHZ+b0Nch60oj239T1M/+5cJ07uH5HmYXKpjQKhBpFD5JmSj69AC1PwuFRfXAxv3BQ/ozg6qQSKym40iAPBRYonadFopZ9fSGr+4a0QHTLVYVft135G1GryzPUlbWM3lamNKAdW5VzrSSxfAi66tZIzvq1rYY3Gh2XXKIsr7QsXyEBhIVDlWlHr4dU2uMOu/pYhQzX+xULDQMKmemqAFD6Dt1OTlDxNqvXrQIKVYqYxTWhah3eJe2efTK8ZTgCSVtfkv2ppm0305tg7PBalrgXQz3JtJeQVyOVyOz0yLQSce+OPvOJuTNZ15vpBZ7jgX/waHdLbnOrQ801GuQRL9yS7XUmnxk5OB1uJe1gf/xVQQvQ0dz2E7cQttf9WXvZb3Xgy7jg9/yh2Dl4UUDGalpXpX7ri26t09tT1RAyR+H9fpX7yFAYVuHHoHvdOvxd9+5n7Ah0r1vSv8lLXC8YHF6Hv8StB367JYnSZkdWktTCbOSsi7lOblGTOsIfW25OHz+OfZ9EXZM5mG4/XneOLDPCniPrOEjFKdq/AoXBU9Wdg/Low4iiGcuJf0Cy+tn+rUe3df4d5V/+Dw==</diagram></mxfile>
