+++ doc/flowDiagram/flowDiagram_getSpellDescription.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2021-12-15T12:04:22.091Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/14.1.8 Chrome/87.0.4280.88 Electron/11.1.1 Safari/537.36" etag="J66KBdyNNctvVpSPpUd1" version="14.1.8" type="device"><diagram id="GPe6hT1JcG3InQYqh8YM" name="Seite-1">7Vrbcts2EP0aPdrDi0xLj7ZsyZ20mbbOJJOnDkSsSMQUQAOgZPXriwV4k0g7cW1FipMXDrEgAew5i90DSoNwsnyYSZKnfwgK2SDw6MMgvBoEwdg7M1c0bJwhOg+cIZGMOpPfGG7Zv1AavdJaMApq60EtRKZZvm2MBecQ6y0bkVKstx9biGx71pwk0DHcxiTrWj8xqlNnHQXnjf0GWJJWM/vR2PUsSfVw6YlKCRXrlim8HoQTKYR2d8uHCWSIXYWLe2/6SG+9MAlcf8sL8Pvsw9Tz0/E7NnvHLv5a3UymJ8HQDbMiWVF6XK5WbyoIpCg4BRzFG4SX65RpuM1JjL1rw7mxpXqZmZZvbpWW4g4mIhPSvh3SaB6dRaZnwbKsZZ9OpxfXV8be9aN0bQVSw0PLVPo1A7EELTfmkbJ3VEK82W6uG8Iir7SlLbKGw9JIyiBJ6pEbHM1NCeUzYPU7qM7YCjAq8EJwP4CKJcs1E9yGpcTATrFX5W6CFEjGeIJ+CEk7rJhwyvHWhGomCv11ZnKQzHgHsnnpz8b0GjT4wTYPdbtFRDDsIWK0Lx6CnuiOMgRL5cQAHyXauh7dF7gNLy8wk0ggupBIhVigM6JAcuJUsBgsS0Q7M1pxFE8BWC91yrApCU/QICEhjCt0luEruWBcYxPuC5LZTGYuPjV70ABw2cxlQyAmSjv+yZxlTCPQS0HZgoE8NfcfUqba4UKwxXFEWCwwFZr142pw8xKKLenypNmiRazVaeN2BYrBeBuXPcfcTkqgBEaLuC+JRPEI5ovXidKhtx2ldUR+LUrH+4rSbg6+EnWErQnXVajccVNAMApMdIbTn4OeMPqOSeSfC61O7i5Pkvv3G7j58ncevv/ck0QuSXxXcTIR3HiGFN1qoqFDiu18okRywWEH69Jk0n/CkRkDpuUDQWSGp4uyY8koxWl6ad6u2o+w0MPVo8TsVtnhuIeYvjIb7IuY8Ei0y4tgPY8OKF56Ye2mo0+u5FEBqlem2L7vpFBeBPbIO6BC6QX77JkKZZIKoSoR2ZIq36hKUDigyFGFTGqFkwvFtNOnwEEmG1t2VFqybfZQkta8N3NO7EyFciKF1VXKyR5zc+7tSJ8e2UIyZSULp2idG8VLOSjlRjfCZdG0CKf2isujTAFRdoHEqh23CLdETYxz+jkaqVFFlU5yqukH1kgv2ie+N/qfIsnfW1qKnlJJHIBWyijwfuPmQGUO4Hi4OmKd9LoU1cLpYLns/JdQsueE6NiE0ugtCCW/VjxHo5TGHVyf/5nnI4ljxl2u+gEUlD8cHZuEqj4Vt2i4FRIHnnTQbur5qo17U+RRNS2dSqgFzRxHsIpmabUH0WRuxMfpGy0suyeSI6j9fveD6iOfSJwrCudNnSaETB3115LXJevwKsB/E8fyzsfsw1cbv3swf3a5Of4CE/jR0RWY7iG9VWA+2xyUEsuEzUEG65gtNiUnFvk3WiuCXU12DLXiyYPiT1wrOmTts1aYZvPTtu1r/T8gvP4P</diagram></mxfile>
