+++ doc/flowDiagram/flowDiagram_invokeSpellAssistantIntent.drawio
