+++ src/response_builders/response_builder_base.py
@@ -1,0 +1,12 @@
+from abc import ABC, abstractmethod
+
+"""
+interface for response builders, that forces the developer to implement get_response correctly
+"""
+
+
+class ResponseBuilderBase(ABC):
+
+    @abstractmethod
+    def get_response(detail: str, casting_level: str) -> dict:
+        pass
